Reorganization of all files
I decided to completely reorganise the layout of the files.  Before I
had kept it exactly the same way the default project had it.  Now I've
reworked it into the way I like it and it also makes it much more
managable.  Also I figured a lot of things out about adding and removing
libraries and such.  And where's a good place to specifie the Resourse
location.